// src/network/model/cartComparisonDto.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * Dolacna.Backend.Api
 * OpenAPI spec version: 1.0
 */
import type { ShopExtendedDto } from './shopExtendedDto';
import type { CartComparisonProductDto } from './cartComparisonProductDto';
import type { CategoryDto } from './categoryDto';

export interface CartComparisonDto {
  shop?: ShopExtendedDto;
  /** @nullable */
  specific_products?: CartComparisonProductDto[] | null;
  /** @nullable */
  missing_products?: CartComparisonProductDto[] | null;
  /** @nullable */
  missing_categories?: CategoryDto[] | null;
  total_price?: number;
}